Add ability to find x tiles by forcing chains
You assume a cell contains a star and follow the chain of implications
this assumption causes. If you reach a contradiction, you know the
initial assumption was false.

findSolution internal functions were extracted and an issue where
changes were being made to the cell data was fixed to allow
findNextStep to use them.